agregando propiedad sueldo al modelo EmpleadoCLS

+++ MiPrimerAplicacion/Models/EmpleadoCLS.cs
@@ -35,7 +35,7 @@ namespace MiPrimerAplicacion.Models
         [Display(Name = "Sueldo")]
         [Range(1,100000,ErrorMessage = "Valor fuera de rango")]
         [Required]
-        public double sueldo { get; set; }
+        public decimal sueldo { get; set; }
 
         [Display(Name = "Tipo Usuario")]
         [Required]
